Move the sectors.c file into the outputs directory
This is the beginning of the outputs/ object. The io_data structure will
be changed to reflect this object.
