+++ src/outputs/sectors.c
@@ -29,7 +29,7 @@
 #include "dislocker/encryption/decrypt.h"
 #include "dislocker/encryption/encrypt.h"
 #include "dislocker/metadata/metadata.h"
-#include "dislocker/sectors.h"
+#include "dislocker/outputs/sectors.h"
 
 
 /*
